- fixed some 64bit type mess
git-svn-id: svn://svn.code.sf.net/p/gwyddion/code/trunk/gwyddion@3367 25697fe4-d644-49b8-9d30-be5465f2adf3

+++ libdraw/gwygradient.c
@@ -1349,7 +1349,7 @@ gwy_gradient_serialize(GObject *obj,
     GwyGradientPoint *pt;
     GArray *points;
     gdouble *data, *rdata, *gdata, *bdata, *adata, *xdata;
-    gsize ndata, i;
+    guint32 ndata, i;
 
     g_return_val_if_fail(GWY_IS_GRADIENT(obj), NULL);
